feat: add SearchCollectionModal
- add mini collection card
- renamed OpenPrivateCollectionModal.tsx to AccessPrivateCollectionModal.tsx
- add useSearchCollection hook
- add getCollectionsByName query

+++ src/lib/utils.ts
@@ -66,7 +66,7 @@ export const getStringDate = (StringDate: string, isRelative = false) => {
 export const getMonthYear = (StringDate: string) => {
   const date = new Date(StringDate);
   return date
-    .toLocaleString(undefined, { month: 'long', year: 'numeric' })
+    .toLocaleString(undefined, { month: 'short', year: 'numeric' })
     .replace(' ', ', ');
 };
